refactoring + states stack

+++ common_lib/src/time.rs
@@ -7,7 +7,7 @@ use std::ops::Sub;
 #[cfg(not(target_arch = "wasm32"))]
 use std::time::Instant as StdInstant;
 
-pub type Duration = u64;
+pub type Duration = f64;
 
 /// A platform-independent replacement for `Instant`
 #[derive(Copy, Clone)]
